+++ api/src/error.rs
@@ -2,6 +2,7 @@ use crate::parser::Error as ParserError;
 
 #[derive(Clone, Debug, thiserror::Error, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
+#[allow(variant_size_differences)]
 pub enum Error {
     #[error("Units are not compatible: {lhs:?}, {rhs:?}")]
     IncompatibleUnitTypes { lhs: String, rhs: String },