+++ api/src/measurement/composable.rs
@@ -1,5 +1,7 @@
-use crate::measurement::Measurement;
-use crate::parser::{Composable, Composition};
+use crate::{
+    measurement::Measurement,
+    parser::{Composable, Composition},
+};
 
 //-----------------------------------------------------------------------------
 // impl Composable
@@ -16,8 +18,10 @@ impl Composable for Measurement {
 //
 #[cfg(test)]
 mod tests {
-    use crate::measurement::Measurement;
-    use crate::parser::{Composable, Composition, Dimension};
+    use crate::{
+        measurement::Measurement,
+        parser::{Composable, Composition, Dimension},
+    };
 
     #[test]
     fn validate_composition() {