+++ api/src/parser/term.rs
@@ -1,8 +1,10 @@
 mod annotation_composable;
 mod composable;
 mod display;
+mod field_eq;
 mod invert;
 mod is_compatible_with;
+mod partial_eq;
 mod reducible;
 mod ucum_unit;
 
@@ -13,11 +15,11 @@ use crate::parser::{Atom, Prefix};
 /// Atom-less Term, which would simple be a Factor (with or without an
 /// annotation) (ex. the 10 in "10" or "10/m" would be an Atom-less Term).
 ///
-#[derive(Clone, Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, Eq)]
 pub struct Term {
-    pub atom: Option<Atom>,
-    pub prefix: Option<Prefix>,
     pub factor: Option<u32>,
+    pub prefix: Option<Prefix>,
+    pub atom: Option<Atom>,
     pub exponent: Option<i32>,
     pub annotation: Option<String>,
 }