+++ api/src/parser/term/ucum_unit.rs
@@ -1,6 +1,5 @@
 use super::Term;
-use crate::reducible::Reducible;
-use crate::ucum_unit::UcumUnit;
+use crate::{reducible::Reducible, ucum_unit::UcumUnit};
 
 impl UcumUnit for Term {
     fn is_special(&self) -> bool {