+++ api/src/unit/field_eq.rs
@@ -1,5 +1,5 @@
-use crate::field_eq::FieldEq;
-use crate::unit::Unit;
+use crate::{field_eq::FieldEq, unit::Unit, Term, UcumUnit};
+use std::cmp::Ordering;
 
 //-----------------------------------------------------------------------------
 // impl FieldEq
@@ -30,14 +30,37 @@ use crate::unit::Unit;
 impl<'a> FieldEq<'a> for Unit {
     #[inline]
     fn field_eq(&self, other: &'a Self) -> bool {
-        self.terms == other.terms
+        fn sort_em(terms: &[Term]) -> Vec<Term> {
+            let mut t = terms.to_vec();
+
+            t.sort_by(|a, b| match a.scalar().partial_cmp(&b.scalar()) {
+                Some(ordering) => ordering,
+                None => Ordering::Equal,
+            });
+
+            t
+        }
+
+        if self.terms.len() != other.terms.len() {
+            return false;
+        }
+
+        let lhs = sort_em(&self.terms);
+        let rhs = sort_em(&other.terms);
+
+        for (l, r) in lhs.into_iter().zip(rhs.into_iter()) {
+            if !l.field_eq(&r) {
+                return false;
+            }
+        }
+
+        true
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::field_eq::FieldEq;
-    use crate::unit::Unit;
+    use crate::{field_eq::FieldEq, unit::Unit};
     use std::str::FromStr;
 
     #[test]
@@ -53,5 +76,9 @@ mod tests {
         let unit = Unit::from_str("100ar").unwrap();
         let other = Unit::from_str("har").unwrap();
         assert!(!unit.field_eq(&other));
+
+        let unit = Unit::from_str("[acr_us].[in_i]").unwrap();
+        let other = Unit::from_str("[in_i].[acr_us]").unwrap();
+        assert!(unit.field_eq(&other));
     }
 }