+++ api/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] 
+
+### Added
+
+- Implemented `Hash` for:
+  - `Classification`
+  - `Composition`
+  - `Prefix`
+- Implemented `Display` for:
+  - `Classification`
+- Implemented `FieldEq` for:
+  - `Term`
+
+### Fixed
+
+- Manually implemeted `PartialEq` for `Term`, comparing the composition and `scalar()`
+  value (like `Measurement` and `Unit`).
+
 ## [0.20.0] - 2021-09-29
 
 ### Changed